+++ public/apps/configuration/panels/permission-list/permission-list.tsx
